♻️ Extract CLI usage helpers and tryDaemon fallback
Move printUsage/printHome into bin/cli/usage.ts and add tryDaemon() for
commands that silently fall back to tmux when the daemon is down (ls,
status, daemon status). Removes duplicate lazy-load logic from lattices.ts.
